refactor: Rename rules.js to validationRules.js and update imports across components
